+++ dserver_search_plugin_mongo/utils_search.py
@@ -7,16 +7,16 @@ import pymongo.errors
 
 from pymongo import MongoClient
 
-from dtool_lookup_server import (
+from dservercore import (
     SearchABC, ValidationError, PaginationParameters, SortParameters)
 
-from dtool_lookup_server.date_utils import (
+from dservercore.date_utils import (
     extract_created_at_as_datetime,
     extract_frozen_at_as_datetime,
 )
 
-from dtool_lookup_server.sql_models import DatasetSchema
-from dtool_lookup_server.schemas import SearchDatasetSchema
+from dservercore.sql_models import DatasetSchema
+from dservercore.schemas import SearchDatasetSchema
 
 from dserver_search_plugin_mongo.config import (
     Config, CONFIG_SECRETS_TO_OBFUSCATE)